+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 0.1.0
 
+- players can catch each other
 - physical world boundaries
 - player look into walk direction
 - integrate matter js as physics engine